requirements.txt
pandas
phidata
python-dotenv
yfinance
packaging
duckduckgo-search
groq
fastapi
uvicorn
python-multipart

sqlalchemy
pgvector
psycopg[binary]
pypdf


google-generativeai